+++ bugrov_av/template/task5/all.h
@@ -2,4 +2,3 @@
 #include "main5.h"
 #include "railway.h"
 #include "cashbox.h"
-
