+++ bugrov_av/template/task6/all.h
@@ -1,2 +1,4 @@
 #pragma once
 #include "main 6.h"
+#include "table.h"
+#include "game.h"
