+++ bugrov_av/template/task6/main 6.h
@@ -6,3 +6,4 @@
 #include <string>
 #include <ctime>
 using namespace std;
+void getxy(int& letter, int& digit);
